WIP dirty hacks to try to enable use OpenGL on macOS
This does NOT work!

OpenGL is fundamentally single threaded, so the real fix is to
delegate all OpenGL (libplacebo) calls to a single thread that
creates a context.

I'm also not sure this is the correct way to create offscreen
contexts, perhaps we should use glfw like others.

Anyway, libplacebo is not very usable with macOS OpenGL 4.2 anyway,
so investing time here is very likely a waste of time.

Signed-off-by: akarin <i@akarin.info>